requirements.txt
# archery
openpyxl
pandoc
weasyprint
# astroimg
astropy
ruamel.yaml
# dosgames
prettytable
# pictures
pillow
# running
apiclient
httplib2
oauth2client
pandas
# weather
aioinflux
